Bug fix attempt with header/source separation.
