+++ libconfluo/confluo/container/sketch/hash_manager.h
@@ -13,16 +13,11 @@ namespace sketch {
 class pairwise_indep_hash {
 
  public:
-  static const size_t PRIME = 39916801UL;
+  static const size_t PRIME;
 
-  pairwise_indep_hash()
-      : pairwise_indep_hash(0, 0) {
-  }
+  pairwise_indep_hash();
 
-  pairwise_indep_hash(size_t a, size_t b)
-      : a_(a),
-        b_(b) {
-  }
+  pairwise_indep_hash(size_t a, size_t b);
 
   template<typename T>
   size_t apply(T elem) const {
@@ -35,40 +30,27 @@ class pairwise_indep_hash {
     return (a_ * elem + b_) % PRIME;
   }
 
-  static pairwise_indep_hash generate_random() {
-    return { utils::rand_utils::rand_uint64(PRIME), utils::rand_utils::rand_uint64(PRIME) };
-  }
+  static pairwise_indep_hash generate_random();
 
  private:
   size_t a_, b_;
 
 };
 
-const size_t pairwise_indep_hash::PRIME;
-
-template<typename T>
 class hash_manager {
  public:
+
   /**
    * Constructor.
    * @param num_hashes number of hashes
    */
-  hash_manager(size_t num_hashes = 0)
-      : hashes_() {
-    this->guarantee_initialized(num_hashes);
-  }
+  explicit hash_manager(size_t num_hashes = 0);
 
   /**
    * Guarantee enough hashes are intialized.
    * @param num_hashes number of hashes
    */
-  void guarantee_initialized(size_t num_hashes) {
-    size_t cur_size = hashes_.size();
-    size_t num_new_hashes = num_hashes > cur_size ? num_hashes - cur_size : 0;
-    for (size_t i = 0; i < num_new_hashes; i++) {
-      hashes_.push_back(pairwise_indep_hash::generate_random());
-    }
-  }
+  void guarantee_initialized(size_t num_hashes);
 
   /**
    * Hash element.
@@ -76,6 +58,7 @@ class hash_manager {
    * @param elem element to hash
    * @return hashed value
    */
+  template<typename T>
   size_t hash(size_t hash_id, T elem) const {
     return hashes_[hash_id].template apply<T>(elem);
   }
